+++ tests/data/src/original-handler.js
@@ -1,0 +1,10 @@
+
+
+exports.handler = function handler(e) {
+    return {
+        statusCode: 200,
+        body: JSON.stringify({
+            message: 'Hello World!',
+        })
+    }
+}