+++ examples/serverless/src/api.js
@@ -1,13 +1,13 @@
 import Dynamodb from 'aws-sdk/clients/dynamodb';
 import SNS from 'aws-sdk/clients/sns';
 import EventBridge from 'aws-sdk/clients/eventbridge';
-import { wrap } from '@baselime/lambda-node-opentelemetry';
 
 const dynamo = new Dynamodb();
 const sns = new SNS();
 const eventbridge = new EventBridge();
 
-export const handler = wrap(async (event) => {
+export const handler = async (event, context) => {
+    context.callbackWaitForEmptyEventLoop = false;
     await sns.publish({ TopicArn: process.env.TOPIC_ARN, Message: 'wow much payload' }).promise()
     await eventbridge.putEvents({
         Entries: [{
@@ -16,24 +16,24 @@ export const handler = wrap(async (event) => {
         }]
     }).promise()
     const random = Math.random();
-    if(random < 0.3) {
-        const response = await dynamo.updateItem({
-            TableName: 'this-table-does-not-exist',
-            ReturnValues: 'ALL_NEW',
-            Key: {
-                id: {
-                    S: 'test'
-                }
-            },
-            UpdateExpression: 'ADD #c :c',
-            ExpressionAttributeNames: {
-                '#c': 'count'
-            },
-            ExpressionAttributeValues: {
-                ':c': { N: '1' }
-            }
-        }).promise()
-    }
+    // if(random < 0.3) {
+    //     const response = await dynamo.updateItem({
+    //         TableName: 'this-table-does-not-exist',
+    //         ReturnValues: 'ALL_NEW',
+    //         Key: {
+    //             id: {
+    //                 S: 'test'
+    //             }
+    //         },
+    //         UpdateExpression: 'ADD #c :c',
+    //         ExpressionAttributeNames: {
+    //             '#c': 'count'
+    //         },
+    //         ExpressionAttributeValues: {
+    //             ':c': { N: '1' }
+    //         }
+    //     }).promise()
+    // }
     const response = await dynamo.updateItem({
         TableName: process.env.DB_NAME,
         ReturnValues: 'ALL_NEW',
@@ -56,4 +56,4 @@ export const handler = wrap(async (event) => {
             data: response
         })
     }
-});
+};