+++ examples/serverless/src/eventbridge.js
@@ -1,8 +1,8 @@
-const { wrap } = require('@baselime/lambda-node-opentelemetry');
 
-exports.handler = wrap(async (event) => {
+exports.handler = async (event, context) => {
+    context.callbackWaitForEmptyEventLoop = false;
     console.log(event)
     return {
       message: 'req processed'
     }
-});
+};