+++ CHANGELOG.md
@@ -14,6 +14,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  The latest layer is: `arn:aws:lambda:${your-region-here}:097948374213:layer:baselime-node:6`
 
 
+ 
+ The latest layer is: `arn:aws:lambda:${your-region-here}:374211872663:layer:baselime-node:11`
+
+ 
+ The latest layer is: `arn:aws:lambda:${your-region-here}:374211872663:layer:baselime-node:12`
+
 ## [0.1.11] 2023-07-22
 
 - publish via CI