+++ CHANGELOG.md
@@ -32,6 +32,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  
  The latest layer is: `arn:aws:lambda:${your-region-here}:097948374213:layer:baselime-node:8`
 
+ 
+ The latest layer is: `arn:aws:lambda:${your-region-here}:097948374213:layer:baselime-node:8`
+
+ 
+ The latest layer is: `arn:aws:lambda:${your-region-here}:374211872663:layer:baselime-node:13`
+
+ 
+ The latest layer is: `arn:aws:lambda:${your-region-here}:374211872663:layer:baselime-node:14`
+
+ 
+ The latest layer is: `arn:aws:lambda:${your-region-here}:374211872663:layer:baselime-node:15`
+
 ## [0.1.11] 2023-07-22
 
 - publish via CI