+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Make sure flushing is not canceled
   
  The latest layer is: `arn:aws:lambda:${your-region-here}:097948374213:layer:baselime-node:9`
+ 
+ The latest layer is: `arn:aws:lambda:${your-region-here}:374211872663:layer:baselime-node:17`
+
+ 
+ The latest layer is: `arn:aws:lambda:${your-region-here}:374211872663:layer:baselime-node:18`
+
 ## [0.1.12] 2023-06-28
 
 - publish via CI